module-4/solid-flask-master/requirements.txt
flask
jinja2
jwcrypto
oic
absl-py
